fix circle dependency in registration
