+++ frontend/src/app/core/components/landing-page/landing-page.component.html
@@ -38,16 +38,7 @@
         <li class="nav-item">
           <button
             type="button"
-            class="btn btn-primary mr-2 mb-2"
-            (click)="signUpRedirect()"
-          >
-            Sign up
-          </button>
-        </li>
-        <li class="nav-item">
-          <button
-            type="button"
-            class="btn btn-outline-primary"
+            class="btn btn-success"
             (click)="signInRedirect()"
           >
             Sign in
