+++ frontend/src/app/core/services/authentication.service.ts
@@ -1,13 +1,10 @@
 import { Injectable } from '@angular/core';
 import { AngularFireAuth } from '@angular/fire/auth';
 import { Router } from '@angular/router';
+import { NewUser } from '@shared/models/user/new-user';
 import { User } from '@shared/models/user/user';
 import * as firebase from 'firebase';
 import { UserService } from './user.service';
-import { NewUser } from '@shared/models/user/new-user';
-import { Providers } from '@shared/models/providers';
-import { NgbModal } from '@ng-bootstrap/ng-bootstrap';
-import { RegistrationDialogComponent } from '@core/components/registration-dialog/registration-dialog.component';
 
 @Injectable({
   providedIn: 'root'
@@ -19,7 +16,6 @@ export class AuthenticationService {
   constructor(
     private angularAuth: AngularFireAuth,
     private userService: UserService,
-    private modalService: NgbModal,
     private router: Router) {
     this.angularAuth.authState.subscribe(user => {
       this.configureAuthState(user);
@@ -32,6 +28,7 @@ export class AuthenticationService {
         this.user = user;
         localStorage.setItem('user', JSON.stringify(this.user));
         localStorage.setItem('jwt', theToken);
+        this.signInWithUid(this.getUIdLocalStorage());
       });
     }
     else {
@@ -49,58 +46,6 @@ export class AuthenticationService {
       });
   }
 
-  doGoogleSignIn(): Promise<void> {
-    const googleProvider = new firebase.auth.GoogleAuthProvider();
-    return this.angularAuth.signInWithPopup(googleProvider).then((auth) => {
-      this.isUidExist(auth);
-    },
-      (error) => console.log("Email exist")
-    );
-  }
-
-  doGithubSignIn(): Promise<void> {
-    const githubProvider = new firebase.auth.GithubAuthProvider();
-    return this.angularAuth.signInWithPopup(githubProvider).then(
-      (auth) => {
-        this.isUidExist(auth);
-      },
-      (error) => console.log("Email exist")
-    );
-  }
-
-  isUidExist(auth: firebase.auth.UserCredential): void {
-    this.userService.login(auth.user.uid)
-      .subscribe((resp) => {
-        if (resp.body !== null) {
-          this.currentUser = resp.body;
-          this.router.navigate(['/portal']);
-        }
-        else {
-          if (auth.credential.providerId === 'google.com') {
-            this.doGoogleSignUp(auth);
-          } else {
-            this.doGithubSignUp(auth);
-          }
-        }
-      });
-  }
-
-  doGoogleSignUp(credential: firebase.auth.UserCredential) {
-    const user = {
-      email: credential.user.email,
-      username: credential.user.displayName,
-      avatarUrl: credential.user.photoURL,
-      firstName: credential.additionalUserInfo.profile[`given_name`],
-      lastName: credential.additionalUserInfo.profile[`family_name`],
-      providerId: Providers.Google,
-      uId: credential.user.uid,
-      providerUrl: credential.credential.providerId
-    } as NewUser;
-
-    const modalRef = this.modalService.open(RegistrationDialogComponent, { backdrop: 'static', keyboard: false });
-    modalRef.componentInstance.details = user;
-  }
-
   registerUser(user: NewUser) {
     this.userService.register(user).subscribe(
       (resp) => {
@@ -108,34 +53,11 @@ export class AuthenticationService {
           this.currentUser = resp.body;
           this.router.navigate(['/portal']);
         } else {
-          user.username = "This Username has already taken";
-          const modalRef = this.modalService.open(RegistrationDialogComponent, { backdrop: 'static', keyboard: false });
-          modalRef.componentInstance.details = user;
-          modalRef.componentInstance.isUsernameTaken = true;
         }
       },
       (error) => {
 
       });
-  }
-
-  doGithubSignUp(credential: firebase.auth.UserCredential) {
-    const user = {
-      email: credential.user.email,
-      username: credential.additionalUserInfo.username,
-      avatarUrl: credential.user.photoURL,
-      providerId: Providers.Github,
-      uId: credential.user.uid,
-      providerUrl: credential.credential.providerId
-    } as NewUser;
-
-    const name: string = credential.additionalUserInfo.profile[`name`];
-    if (name != null) {
-      [user.firstName, user.lastName = ''] = name.split(' ');
-    }
-
-    const modalRef = this.modalService.open(RegistrationDialogComponent, { backdrop: 'static', keyboard: false });
-    modalRef.componentInstance.details = user;
   }
 
   logout(): Promise<void> {
@@ -159,6 +81,10 @@ export class AuthenticationService {
 
   getUser(): User {
     return this.currentUser;
+  }
+
+  setUser(user: User): void {
+    this.currentUser = user;
   }
 
   getUIdLocalStorage(): string {
